Delete getting_started.rst (#44)
This file was added in #26, but I am not really sure that it belongs in this repository.

It is duplication of instructions that are in Qiskit, but it has not been
updated since it was added and is falling out of date.

See. e.g. the update at https://github.com/Qiskit/qiskit/commit/be8b9502c6d0ae201d9aad542a9ac9123a675a19
which is not included in this repository.