Sync jinja sphinx theme templates from qiskit metapackage (#49)
This commit syncs the jinja2 templates for the the sphinx layout and
version/language table templates from the version in the metapackage.
The metapackage has continued to develop these layout files
independently to improve the layout of the hosted documentation and this
commit syncs those local changes back to the theme as a precursor to
unifying the theme across projects.

At the same time the custom sphinx extension versionutils which is used
to extract a version list for the stable version list. Some work will be
needed to manage the sphinx extension interface to enable configuring
the use of previous versions and the translation list. Right now these
are hard coded in the plugin code, but before the sphinx extension can
be used these will need to be configurable by a user and documented.