Fix deprecation due to `package_data` in `setup.py` (#212)
When building `qiskit_sphinx_theme`, we would get this deprecation:

```
    ############################
    # Package would be ignored #
    ############################
    Python recognizes 'qiskit_sphinx_theme.static.css' as an importable package,
    but it is not listed in the `packages` configuration of setuptools.

    'qiskit_sphinx_theme.static.css' has been automatically added to the distribution only
    because it may contain data files, but this behavior is likely to change
    in future versions of setuptools (and therefore is considered deprecated).

    Please make sure that 'qiskit_sphinx_theme.static.css' is included as a package by using
    the `packages` configuration field or the proper discovery methods
    (for example by using `find_namespace_packages(...)`/`find_namespace:`
    instead of `find_packages(...)`/`find:`).

    You can read more about "package discovery" and "data files" on setuptools
    documentation page.
```

As suggested by both this warning and
https://setuptools.pypa.io/en/latest/userguide/datafiles.html#include-package-data,
the solution is that we should be setting the `packages` argument to
list the `static` folders.

Because we have `include_package_data=True` and those non-Python files
listed in `MANIFEST.in`, they get included in the final wheel. No need
for the `package_data` keyword, which was redundant with our
`MANIFEST.in`.

I confirmed this works by using `python -m build` and comparing the
`qiskit_sphinx_theme-1.11.0rc1.dist-info/RECORD` entries before and
after, which lists every file included in the wheel.

--

This also adds a `pyproject.toml` declaring our `[build-system]`. This
is heavily encouraged by modern Python packaging.

And it removes the `zip_safe` arg, which setuptools calls "obsolete":
https://setuptools.pypa.io/en/latest/deprecated/zip_safe.html