Remove unnecessary top-level `qiskit_sphinx_theme.py` (#213)
I'm not totally sure why this was added in
https://github.com/Qiskit/qiskit_sphinx_theme/pull/8/files. It is not
necessary to fix search, which I believe was fixed by the change in
`layout.html`.

This code was unused. I added a `raise ValueError()` inside `get_path()`
and that never got executed when using `tox -e py -r -- -E`.

Generally, Sphinx's HTML Theme extension would not be looking for a
Python file: it expects CSS, JS, images, and `theme.conf`.