Reorganize to allow theme variants (#216)
## Summary

We plan to add variants of the Sphinx theme so that we can differentiate
between Core vs the Ecosystem.

They'll be distributed in the same package, and users only need to
choose the `html_theme` in their `conf.py`. They will also share a
"base" theme and override the CSS/images/HTML templates they care about.
See https://github.com/Qiskit/qiskit_sphinx_theme/pull/215 for an
example of this.

This PR reorganizes our folder structure so that we can add variants in
followup PRs.

## Details

We call the folder `pytorch_base` because we likely will want
Pytorch-based variants for Core vs Ecosystem, and they'll use this
folder as their base. Those variants will be added in follow ups.

This also simplifies our `MANIFEST.in` and `setup.py` with
@jakelishman's recommendation from
https://github.com/Qiskit/qiskit_sphinx_theme/pull/212#pullrequestreview-1344320986.
The benefit of that is we make fewer changes when adding new theme
variants (we have to add `graft qiskit_sphinx_theme/<variant>/static` to
MANIFEST.in). I verified this works by using `python -m build` and
inspecting the RECORD of the wheel, which shows the contents.