Remove unused `layout_old.html` (#255)
We now only use `layout.html` in modern versions of the Qiskit Sphinx
Theme.

Keeping around the dead code makes our repository more confusing. It's
not clear to the lay reader why that file exists.

We have version control if we ever want to see what the code was.

This will have no impact on users because the template is not used. And
this change is only included in the theme 1.12+ -- prior versions will
still have this file.

--

We also remove the example docs overridden template of `versions.html`.
That file is no longer used. And it is no longer necessary to override
the template - you only need to fill in the data.