Include `sphinxcontrib.jquery` in the project dependencies & fix other JavaScript issues (Cherry-pick of #277) (#278)
Sphinx 6 stopped including jQuery by default, but we're still using it
in some places. As a result, the docs' functionality is broken in some
places like the language selection not being clickable:
https://github.com/Qiskit/qiskit_sphinx_theme/issues/272

I wanted to remove jQuery via
https://github.com/Qiskit/qiskit_sphinx_theme/issues/275, but it is too
hard to safely due with Pytorch because we have 105 usages. We will need
to wait for the switch to Furo.

So, for now, we will include jQuery by default. We also try to activate
the extension by default with the line
`app.setup_extension("sphinxcontrib.jquery"),` but I'm having issues
with that actually working. So, projects may need to still add
`"sphinxcontrib.jquery"` explicitly in their extensions in `conf.py` -
but they at least won't have to explicitly include
`sphinxcontrib-jquery` in their requirements.

--

This PR also fixes other issues with our JavaScript, as reported by the
developer console:

1. `languages.html` was not recognizing jQuery. So, rewrite it to modern
JavaScript
2. `pytorchAnchors.bind();` was not defined so caused some of our setup
to fail
3. `handleLeftMenu:` function was failing because
`document.getElementById("docs-tutorials-resources")` does not exist.
Since that function failed, it looks like it wasn't actually necessary
anymore.