Set up infrastructure for Jest (JavaScript) tests (#282)
Prework for https://github.com/Qiskit/qiskit_sphinx_theme/issues/208.

Abby and I decided that we want Jest tests for two purposes:

1. Checking that web components are in sync between the Pytorch and Furo
themes, e.g. we didn't forget to update one of the folders.
1. We could do this either by using Snapshot testing or by simply
checking that the file contents are equal.
2. We have some non-trivial logic in
https://github.com/Qiskit/qiskit_sphinx_theme/pull/267 and want to write
automated unit tests.

This PR adds the basic infrastructure, like setting up CI to install
Node.js. That will make it easier to review follow-up PRs that add
actual tests.

> We could do this either by using snapshot testing

I think we'd benefit from adding snapshot testing for the top nav bar
rendering correctly. I want to do this in a follow up PR so we can
properly decide if we want to add snapshots or not; the decision to use
snapshots is separate from whether to use Jest generally.