Add experimental `_qiskit_furo` theme (#269)
This PR adds a private `_qiskit_furo` theme to `qiskit_sphinx_theme`,
using Sphinx's inheritance mechanism. That allows us to build
`example_docs` with Furo by using `THEME=_qiskit_furo tox -e py`.

The Furo theme still has some CSS issues, documented in
https://github.com/Qiskit/qiskit_sphinx_theme/issues/232. But I'd like
to land this theme and make those improvements in followup PRs. That
will make it easier for people (like Digital) to collaborate on
improving this theme, and it will make reviews more focused on the
specific CSS fix.

> using Sphinx's inheritance mechanism

This is much more maintainable than vendoring all the Furo code in our
repository.

For HTML template overrides, we add the comments `<!-- QISKIT CHANGE:
start. ... -->` and `<!-- QISKIT CHANGE: end. -->` to make clear our
changes. We also put substantial changes in dedicated HTML Jinja
template files, then use `{% include "..." %}`. That will make it easier
to sync with upstream Furo.

## Web components

We duplicate the same top-nav-bar.js file for both Pytorch and Furo.
This is because Sphinx requires the file to be in the `static/` folder
for both themes. We can't put it in a common folder. Using symlinks to
copy the file in didn't work either due to Python packaging.

We use a Jest test to ensure that the two files are in sync.