Fix Furo side bars & improve footer styling (#294)
Before, the left table of contents and the right page of contents would
not render properly on mobile. They would not show the whole content for
example. This is because they were not accounting for our top nav bar.

--

This PR also fixes the "Was this page helpful?" question not properly
having lines above and below it.
<img width="841" alt="Screenshot 2023-05-03 at 8 56 29 AM"
src="https://user-images.githubusercontent.com/14852634/235955023-16bba857-86cb-4620-9c46-471707e8331e.png">

--

Finally, this improves mobile to not show the project name at the top of
the page. We don't render it for desktop, and don't want to for mobile.

<img width="700" alt="Screenshot 2023-05-03 at 8 56 54 AM"
src="https://user-images.githubusercontent.com/14852634/235955160-8ffcc6e5-6e77-4c76-b5ee-ed600e577d66.png">