Users should set `docs_url_prefix` for Translations and Previous Releases (#321)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/303.

Everyone's local copy of `versionutils.py` already had the option
`content_prefix`, which had to be set to e.g. `ecosystem/finance` or
`documentation`. That was already being used by the Translations
feature. (But it was only for their local projects; qiskit_sphinx_theme
never had this option.)

This PR makes some improvements to `content_prefix`:

* Renames to the more clear `docs_url_prefix`
* Eagerly errors if it is not set and either Translations or Previous
Releases are used
* Refactors Previous Releases logic to use this option.
* This results in simpler logic, and the URL will now be determined by
Sphinx at build time, rather than by dynamic JavaScript code at runtime.
That is valuable because it makes our `<a href>` an actual link, which
better complies with the Semantic Web.