Add Furo to install_requires and bump Sphinx requirement to 6.0+ (#339)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/330.

While we might not be ready to fully switch all production users to Furo
in Theme 1.13, we at least will have made enough progress that it is
ready for them to try it out. We want to make that setup easier.

The main change that impacts us from the prior release
(https://github.com/pradyunsg/furo/commits/2023.05.20) is that Furo now
requires Sphinx 6.0+. Note that Sphinx 6 drops Python 3.7. Most Qiskit
projects have already made this change, but not all yet. So we probably
need to coordinate this PR with those projects.