Move web component test to Python (#358)
When we added our tests for web components, we only had Jest/JavaScript
test infrastructure. No Pytest/Python test infrastructure. So, I wrote
the tests in JavaScript.

But the tests can easily be written in Python instead.

Why make this change? If we land
https://github.com/Qiskit/qiskit_sphinx_theme/pull/345, then it will
require now using Docker to run the tests locally (but you can use CI to
get the updated snapshot files). So, I want to have our JavaScript tests
only be things that truly require JavaScript: our visual regression
tests, and possibly in the future tests of our JavaScript code.

If we decide to not land
https://github.com/Qiskit/qiskit_sphinx_theme/pull/345, then we may want
to remove this JavaScript test infrastructure entirely.