Make Furo main content responsive, rather than left side bar (#386)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/360.

Now the main content will grow bigger as the site expands, rather than
the left side bar. That uses screen real estate better. The left side
bar will always be the same size of 18em (rather than default of 15em),
or hidden on mobile. The right side bar will also always be 15em, or
hidden by an icon.

However, we set the max width for the main content to 60em. As explained
in
https://www.mediawiki.org/wiki/Reading/Web/Desktop_Improvements/Features/Limiting_content_width#Goals_and_Motivation,
it is useful to limit content width for prose-heavy websites because too
long of lines are harder to read. With Furo's font size, a max-width of
60 em results in lines with up to ~125 characters.

Once the max width is hit, the site basically stays fixed. That is,
whitespace gets added to the right of the page table of contents and it
stays fixed in position. That was much easier to implement and I think
is a better UX because it keeps the sidebar closer to page contents.

![Screenshot 2023-06-16 at 8 49 20
AM](https://github.com/Qiskit/qiskit_sphinx_theme/assets/14852634/eea3ec07-34ee-401c-9d27-7a5b8b8fca24)

![Screenshot 2023-06-16 at 9 34 07
AM](https://github.com/Qiskit/qiskit_sphinx_theme/assets/14852634/3b49d294-a605-459e-8eae-e855924b4c37)

This also simplifies the fix for setting the color of links in the table
of contents to use CSS variables.