Add Previous Releases feature to Furo (#384)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/381.

The implementation mostly reuses our code for translations
(https://github.com/Qiskit/qiskit_sphinx_theme/pull/380).

See https://github.com/Qiskit/qiskit_sphinx_theme/pull/355 for an
explanation of why we diverge Furo's standard approach of the top-level
entry being clickable and going to a dedicated HTML page. Instead, we
have the entry look like a sidebar entry, and still have the entire div
be clickable.

## Reduces padding for translations

This reduces padding for the translations header, which aligns it with
the rest of the sidebar entries. I'm not sure why I had set up so much
padding before.

@SooluThomas and I thought this is a cleaner design, and that it takes
less screen real estate. But we can add back the padding if people want.

## Tweaks example_docs table of contents

To make our example docs more clear, this adds a new Special Elements
section. On Pytorch, that results in the ToC looking like this:


![pytorch-sidebar](https://github.com/Qiskit/qiskit_sphinx_theme/assets/14852634/8faeb393-35a6-4bc5-8868-15e0082c4030)