Migrate from setup.py to pyproject.toml (#422)
`pyproject.toml` is now the standardized way to set packaging metadata
for Python packages. When possible, setuptools recommends using it
rather than a dynamic `setup.py` time.

Switching to `pyproject.toml` specifically unblocks
https://github.com/Qiskit/qiskit_sphinx_theme/pull/419.

Another benefit is we can deduplicate setting the version in two places.

## Gets rid of "universal" wheel

We were incorrectly claiming in `setup.cfg` that the package is
"universal", which means it works with both Python 2 and Python 3. That
is a lie. We don't support Python 2.