Add Furo styling to `qiskit-call-to-action-item` (#434)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/327.

This uses an extremely similar style to Pytorch:


![pytorch-cats](https://github.com/Qiskit/qiskit_sphinx_theme/assets/14852634/8df4e519-b09d-4cac-ae88-23e2fa693a72)

It's main change is using our color variables. The padding is also a
little less intense.

The icon implementation is awkward, but I couldn't figure out how to
reuse the icon from `icons.html` easily because we still need to support
Pytorch. So, I went with this implementation and opened
https://github.com/Qiskit/qiskit_sphinx_theme/issues/433 to improve it
once we drop Pytorch.