Fix Furo page layout on medium-sized screens (Cherry-pick of #450) (#454)
We changed Furo's left sidebar from `15em` to `18em`, but never fixed
the `content` to go from `46em` to `43em`.

This messed up the layout on medium-sized screen, right before the
tablet mode kicks in with the right sidebar hiding. The page would try
to horizontally scroll, which doesn't work properly.