Add back Furo's highlighting of page ToC (Cherry-pick of #455) (#457)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/368.

I realized that we cannot fix this upstream. We always need to fork
`furo.js` because the `3.5rem` for our qiskit-top-nav-bar needs to be
hardcoded. It doesn't update Furo's `header.top` like I was hoping.

Thanks to now using Webpack and `sphinx-theme-builder`, we can vendor
the two JavaScript files without issue.