Add copyright headers along with Furo license (Cherry-pick of #464) (#467)
## Adds Furo license

It's extra safe to do this because we vendor about 6 files from Furo. 

We use the prior art of Pytorch of including the license directly in
`LICENSE` at the bottom. Pytorch is also an MIT license and we
substantially vendored it, so it's a similar setup.

## Fixes LICENSE inclusion

The switch to `pyproject.toml` accidentally changed us including the
LICENSE file to instead have the text `Apache 2`.

## Adds copyright file headers

We should have been doing this from the start because Qiskit projects
are expected to put copyright headers on every file. This becomes
especially useful if we more substantially vendor Furo code in
https://github.com/Qiskit/qiskit_sphinx_theme/pull/463 to make clear
what is what.

All files that are originally written by Qiskit contributors use the
standard Apache 2 license we already use in this project.

Meanwhile, files that are vendored from Furo now use this standard
notice:

> This file is vendored from Furo (created by Pradyun Gedam) and used
under the MIT license.

Even though the MIT license does not require file headers, and Furo
itself does not use it, it is a good practice to make clear what
originally comes from Furo.

## Also renames `search_sidebar.html`

This file is vendored from Furo, but we had it under
`custom_templates/search_sidebar.html` rather than the original
`sidebar/search.html`. This PR renames the file so that we more properly
attribute Furo and the MIT license.