Rename `npm test-snapshots` to `npm test-qiskit` (#473)
Prework for adding visual regression testing to the ecosystem theme in
https://github.com/Qiskit/qiskit_sphinx_theme/issues/470.

`npm run test-snapshots` is only run locally. I know that I'll want to
be able to easily run only the Qiskit theme or only the Ecosystem theme,
since the snapshots only are for what is unique to each theme. Snapshot
tests are slow, so I want to run the minimum amount of tests. We'll have
`npm run test-ecosystem` in a followup.

This also makes `npm run test-qiskit` more useful by first running `tox
-e qiskit`. That reduces the risk of forgetting to rebuild the docs
beforehand.