Ecosystem: Add project name to mobile header (#494)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/484.

For the Ecosystem, we revert back to Furo's default mobile header
design, which includes the project name. That works well because we
don't have the top Qiskit nav bar.

Our only change is to use a Carbon icon for the hamburger icon.

### Vendors `page.html`

Rather than inherting `page.html` from our base Qiskit theme, we
re-vendor `page.html` from Furo directly for the ecosystem. That is
because the Qiskit theme makes several changes that we don't like, like
adding the top Qiskit nav bar.

Our only divergence from Furo with the Ecosystem is to add analytics in
the footer.

### Removes `docs` text from Search

This change impacts both the Qiskit and Ecosystem themes. For most
projects, it is redundant to say "docs": the only site for the entire
project is the docs site! So search is for the entire project.

The only exception is Terra, since search only applies to the docs
subsite and not qiskit.org. But we can workaround that by changing the
project name in Terra's `conf.py` to say `Qiskit documentation`.