Ecosystem: show project name and logo in left sidebar (#505)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/483.

Users can set a custom logo with `html_logo` or `light_logo` and
`dark_logo` from `html_theme_options`, like you'd normally do with Furo.
Otherwise, we'll use the Qiskit Ecosystem logo by default.

Projects that don't want a logo can set `disable_ecosystem_logo` in
`html_theme_options`. They also can disable the project text with
`sidebar_hide_name` in `html_theme_options`.

## Ecosystem logo is clickable

By default, Furo makes the entire Brand section clickable to go to the
home page. We keep that when you're not using the Ecosystem logo.

But with the Ecosystem logo, we make the image itself clickable to
https://qiskit.org/ecosystem/. This closes
https://github.com/Qiskit/qiskit_sphinx_theme/issues/485. It's our "Back
to Qiskit.org" mechanism. The site title is not clickable, only the
logo.

## Reverts search placeholder

In Qiskit, we show the project name in the Search box, e.g. `Search
Qiskit Terra documentation`. That is to make clear that you're not
searching the entire site.

But for ecosystem projects, their website usually is the entire project!
And we already show the project name right above the search. So the
placeholder is noisy and redundant for the ecosystem.

(Qiskit theme is unchanged.)