Add tests for Ecosystem removing top nav bar (#502)
The testing strategy is to get the exact page position of these elements
and make sure they're where we expect, i.e. that we're not adding the
`3.5rem` like before. We don't take screenshots because they would be
sensitive to unrelated visual changes.

Otherwise, the tests are almost identical to what we have for Qiskit.