Fix test for left side bars being covered by Qiskit top nav bar (#481)
https://github.com/Qiskit/qiskit_sphinx_theme/pull/412 moved
translations from the top-most element of the left sidebar to the
bottom-most element.

But I forgot to update some of our tests. They need to check that our
Qiskit top nav bar is not covering the searchbox, since that is now our
top-most element in the sidebar.

This also runs Prettier on the tests, hence some extra commas being
added.