Ecosystem: change color scheme back to purple (#509)
We decided in the docs squad that the Ecosystem Theme is distinct enough
to not need a new color scheme. We all like purple more than magenta.

But, IBM specific projects like Qiskit IBM Runtime will want to use
blue. So, we still define blue and allow setting it.

We use blue in the example docs because we want our visual regression
tests to make sure customizing the color still works.