Turn off CI for pushes (#565)
We use GitHub branch protection to require opening a PR for `main` and
the `1.*` branches; you cannot push directly. For `main`, we also
require that CI runs against the latest `main`.

So, there is no need to run CI again for pushes since it already runs
with PRs. Don't waste compute resources, which have a real environmental
cost.