Remove Qiskit top nav bar (#581)
Part of https://github.com/Qiskit/qiskit_sphinx_theme/issues/577.

This is great because it allows us to remove the custom web components
mechanism. We could always add that back in the future if necessary, but
it does add a lot of complexity. So removing it makes this repo more
sustainable.

We can also remove needing to replace `furo.js` with
`qiskit-sphinx-theme.js`.