Remove custom mobile header & search box (#582)
Part of https://github.com/Qiskit/qiskit_sphinx_theme/issues/577.

We used to have a custom mobile header for the Qiskit theme to account
for the Qiskit.org top nav bar. The Ecosystem theme meanwhile reverted
to the default Furo header. We now always use the default Furo header.

This also reverts a change to the placeholder text in the search box to
mention the project name. That was to make more clear that search was
only for Qiskit docs and not all of qiskit.org. That's not relevant now.