Turn off live preview in PRs (#600)
By turning this off, we'll be able to archive
https://github.com/Qiskit/gh-actions and inline it to live in the only
remaining usage at qiskit/documentation. That simplifies our infra.

We're not very actively developing this repository, so the live previews
are less important than before. We can still easily preview docs by
looking at the GitHub artifact. It's not crucial to be able to use the
GitHub deployment instead.