Use standard chevron icon for call to actions (#602)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/433. This is
mainly to reduce tech debt.

Note that the color changed to a darker purple. Our color palette in the
Furo theme only has the darker color of purple available, so this is
acceptable.

The icon is also a little smaller. I didn't spend much time trying to
get the original size because the new size seems fine and this custom
directive is not widely used.