Prepare 2.0.0rc1 (#617)
Closes https://github.com/Qiskit/qiskit_sphinx_theme/issues/616.

We decided to not change of the option names in 2.0 because it is too
disruptive. We also don't remove `disable_ecosystem_logo`, even though
it no longer does anything.

This PR refreshes the README to better explain how to use the docs with
recent changes, such as GitHub Pages rather than qiskit.org.