Fix CI (#627)
Our Docker image stopped working with Tox. 

We also were getting failures from Sphinx with our footnotes in
paragraph.rst. I suspect this is because of new versions of Sphinx
probably adding the warning. That file is solely used for our testing,
rather than our docs guide, so the simplest fix was to delete the
problematic lines.

Also fixes CVEs.